Fix camera controller and bug where one enemy got destroied by a wall
Fix Cinemachine:
The project has been created using Cinemachine to make the camera follow the player. But its currently broken. Fix the camera by considering what you need to do with the Cinemachine package.

Vanishing Enemy:
One of the enemies is disappearing without being captured. Figure out why and fix it so that this bug is not happening.