Stop the particles after X amount of time
Stop The Party:
The celebration particles should last for 1 second then stop. But they keep looping. Use our bool variable to fix the celebration so that it doesn't keep playing.

Used a timer that start counting after the player collected the last enemy. if the timer is greater or equal to the wait time we then stop the particles