Tricky Patterns
create more interesting (and varied) patterns for their dodging.

Created SpawnCubesAtRandom, that finds all the enemies in level and set it to a list, and also an array of spawn points

Created 2 mehods that returns random spawn point and set a gameobject to false.

if the timer reaced the wait time we instantiate new enemies at random positions