Timer
The timer is now counting up, and reset as soon as the player collides with all the enemies and the scene reloads
Counting Up:
Finish the Timer.cs script so that we display an ongoing timer in the top corner of the screen. For bonus points, show the timer with no decimal places, just whole numbers.

Restart The Action:
Restart the level after the player collects all the enemies and the celebration fireworks have finished. Make sure your timer resets when the level restarts.