Particles for colliding with enemy & Bullet time
Gotcha:
Create a new particle effect and trigger it to play when an enemy is captured.

Bullet Time:
When the Hero captures an enemy, briefly slow the game down for just a short amount of time (half a second or so). You know, bullet-time style. For bonus points, give the player some sort of indication that the game has briefly slowed down.